+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Breaking changes are marked with ⚠️.
 **Fixed**
 
 - Fix bug where `route().current()` could incorrectly return `true` on URLs with no parameters ([#377](https://github.com/tighten/ziggy/pull/377))
+- Fix several other bugs in `route().current()` with params ([#379](https://github.com/tighten/ziggy/pull/379))
 
 ## [v1.0.3] - 2020-11-20